# internal/adapters/http/docs/swagger.yaml
basePath: /api
definitions:
...
  response.DataResponse-user_BatchCreateResultDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/user.BatchCreateResultDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-user_UserWithRolesDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/user.UserWithRolesDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.ErrorDetail:
    properties:
      code:
        description: 业务错误码（小写下划线）
        type: string
      details:
        description: 额外详情（如验证错误列表）
      message:
        description: 错误消息
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
    type: object
  response.MessageResponse:
    description: 纯消息响应格式
    properties:
      code:
        description: HTTP 状态码
        type: integer
      message:
        description: 消息描述
        type: string
    type: object
...
  response.PagedResponse-user_UserWithRolesDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        description: 数据列表
        items:
          $ref: '#/definitions/user.UserWithRolesDTO'
        type: array
      message:
        description: 消息描述
        type: string
      meta:
        allOf:
        - $ref: '#/definitions/response.PaginationMeta'
        description: 分页信息
    type: object
  response.PaginationMeta:
    properties:
      has_more:
        description: 是否有下一页
        type: boolean
      page:
        description: 当前页码
        type: integer
      per_page:
        description: 每页数量
        type: integer
      total:
        description: 总记录数
        type: integer
      total_pages:
        description: 总页数
        type: integer
      warning:
        description: 页码越界警告
        type: string
    type: object
...
  user.AssignRolesDTO:
    properties:
      role_ids:
        items:
          type: integer
        type: array
    required:
    - role_ids
    type: object
  user.BatchCreateDTO:
    properties:
      users:
        items:
          $ref: '#/definitions/user.BatchItemDTO'
        maxItems: 100
        minItems: 1
        type: array
    required:
    - users
    type: object
  user.BatchCreateErrorDTO:
    properties:
      email:
        type: string
      error:
        type: string
      index:
        type: integer
      username:
        type: string
    type: object
  user.BatchCreateResultDTO:
    properties:
      errors:
        items:
          $ref: '#/definitions/user.BatchCreateErrorDTO'
        type: array
      failed:
        type: integer
      success:
        type: integer
      total:
        type: integer
    type: object
  user.BatchItemDTO:
    properties:
      email:
        type: string
      nickname:
        maxLength: 50
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      real_name:
        maxLength: 100
        type: string
      role_ids:
        items:
          type: integer
        type: array
      signature:
        maxLength: 255
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  user.CreateDTO:
    properties:
      email:
        type: string
      nickname:
        maxLength: 50
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      real_name:
        maxLength: 100
        type: string
      role_ids:
        items:
          type: integer
        type: array
      signature:
        maxLength: 255
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  user.RoleDTO:
    properties:
      description:
        type: string
      display_name:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  user.UpdateDTO:
    properties:
      avatar:
        maxLength: 255
        type: string
      bio:
        type: string
      email:
        type: string
      nickname:
        maxLength: 50
        type: string
      phone:
        type: string
      real_name:
        maxLength: 100
        type: string
      signature:
        maxLength: 255
        type: string
      status:
        enum:
        - active
        - inactive
        - banned
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    type: object
  user.UserWithRolesDTO:
    properties:
      avatar:
        type: string
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      nickname:
        type: string
      phone:
        type: string
      real_name:
        type: string
      roles:
        items:
          $ref: '#/definitions/user.RoleDTO'
        type: array
      signature:
        type: string
      status:
        type: string
      type:
        description: '"human" | "service" | "system"'
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:40012
info:
  contact:
    email: lwmacct@icloud.com
    name: API Support
    url: https://github.com/lwmacct/251117-go-ddd-template/issues
  description: 基于 DDD+CQRS 架构的 Go Web 应用 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/lwmacct/251117-go-ddd-template
  title: Go DDD Template API
  version: 1.0.3
paths:
...
  /api/admin/users:
    get:
      consumes:
      - application/json
      description: 分页获取所有用户列表（包含角色信息）
      parameters:
      - default: 20
        description: Limit 每页数量，默认 20，最大 1000
        in: query
        maximum: 1000
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: Page 页码，从 1 开始
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Search 搜索关键词（用户名或邮箱）
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 用户列表
          schema:
            $ref: '#/definitions/response.PagedResponse-user_UserWithRolesDTO'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 用户列表
      tags:
      - Admin - User Management
    post:
      consumes:
      - application/json
      description: 管理员创建新用户账号，可同时分配角色
      parameters:
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateDTO'
      produces:
      - application/json
      responses:
        "201":
          description: 用户创建成功
          schema:
            $ref: '#/definitions/response.DataResponse-user_UserWithRolesDTO'
        "400":
          description: 参数错误或用户名/邮箱已存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建用户
      tags:
      - Admin - User Management
  /api/admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: 管理员删除指定用户（物理删除或软删除）
      parameters:
      - description: 用户ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 用户删除成功
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "400":
          description: 无效的用户ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - Admin - User Management
    get:
      consumes:
      - application/json
      description: 根据用户ID获取用户详细信息（包含角色信息）
      parameters:
      - description: 用户ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 用户详情
          schema:
            $ref: '#/definitions/response.DataResponse-user_UserWithRolesDTO'
        "400":
          description: 无效的用户ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 用户详情
      tags:
      - Admin - User Management
    put:
      consumes:
      - application/json
      description: 管理员更新用户的基本信息和状态
      parameters:
      - description: 用户ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 用户更新成功
          schema:
            $ref: '#/definitions/response.DataResponse-user_UserWithRolesDTO'
        "400":
          description: 无效的用户ID或参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新用户
      tags:
      - Admin - User Management
  /api/admin/users/{id}/roles:
    put:
      consumes:
      - application/json
      description: 管理员为指定用户分配角色（会覆盖现有角色）
      parameters:
      - description: 用户ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: 角色ID列表
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.AssignRolesDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 角色分配成功
          schema:
            $ref: '#/definitions/response.DataResponse-user_UserWithRolesDTO'
        "400":
          description: 无效的用户ID或参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 用户不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 分配角色
      tags:
      - Admin - User Management
  /api/admin/users/batch:
    post:
      consumes:
      - application/json
      description: 管理员从 CSV 等来源批量创建用户，支持部分失败（单个失败不影响其他用户）
      parameters:
      - description: 用户列表（最多 100 个）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.BatchCreateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: 批量创建结果
          schema:
            $ref: '#/definitions/response.DataResponse-user_BatchCreateResultDTO'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 批量创建用户
      tags:
      - Admin - User Management
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"